agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "owaisyosuf", email = "owais.251990@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.4.400",
    "openai-agents>=0.0.9",
]

[project.scripts]
agent = "agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
